+++ src/apiflask/__init__.py
@@ -1,3 +1,7 @@
+from __future__ import annotations
+
+import typing as t
+
 from . import fields as fields
 from . import validators as validators
 from .app import APIFlask as APIFlask
@@ -13,4 +17,19 @@ from .schemas import Schema as Schema
 from .security import HTTPBasicAuth as HTTPBasicAuth
 from .security import HTTPTokenAuth as HTTPTokenAuth
 
-__version__ = '2.1.2-dev'
+
+def __getattr__(name: str) -> t.Any:
+    if name == "__version__":
+        import importlib.metadata
+        import warnings
+
+        warnings.warn(
+            "The '__version__' attribute is deprecated and will be removed in"
+            " APIFlask 2.1.2. Use feature detection or"
+            " 'importlib.metadata.version(\"apiflask\")' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return importlib.metadata.version("apiflask")
+
+    raise AttributeError(name)